+++ src/index.html
@@ -16,8 +16,6 @@
   <body>
     <include src="./partials/preloader.html"></include>
     <include src="./partials/modal.html"></include>
-
-    <button id="btnLogout" type="button">Log out</button>
     <include src="./partials/header.html"></include>
     <include src="./partials/gallery.html"></include>
     <include src="./partials/pagination.html"></include>
