+++ src/index.js
@@ -7,4 +7,4 @@ import './js/api-service';
 import './js/modal';
 import './js/search-film';
 import './js/header';
-import './js/authentication/authentication';
+import './js/authentication';
