+++ src/partials/authentication.html
@@ -1,3 +1,9 @@
 <div class="auth">
-  <div class="auth__form"></div>
+  <div class="auth__window">
+    <h2 class="auth__heading">To access the library, please, log in.</h2>
+    <div class="auth__form"></div>
+    <button type="button" class="auth__cancel" data-action="cancel">
+      Cancel
+    </button>
+  </div>
 </div>
