+++ src/sass/components/_authentication.scss
@@ -4,5 +4,24 @@
   left: 0;
   width: 100vw;
   height: 100vh;
+  background-color: #000000a0;
+}
+
+.auth__cancel {
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.auth__heading {
+  text-align: center;
+}
+
+.auth__window {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  padding: 40px 20px;
   background-color: #eeeeee;
 }
